update for new home az contract from nma (#3151)
* update for new home az contract

* fix lint

* address comment

* fix lint

* address comment

* Add "AppliedFixes" to HomeAzResponse

The NMAgent team has added an API version field to the HomeAZ response
as a means to indicate whether or not certain bugfixes have been
applied. Since this API Version field conveys no meaningful information,
this adds an "AppliedFixes" field to the response to make it meaningful
to users.

* Implement fmt.Stringer for HomeAZFix

This is to ensure that these are prett-printed properly for clients that
log them.

* Use nmagent.(HomeAZResponse).ContainsFixes in CNS

This alters the response to use the ContainsFixes method added to the
HomeAzResponse. Rather than using API Version directly, the more
semantically-meaningful constants from the nmagent package are used
instead to indicate whether or not some bugfixes have been applied.

* add nma client test

* fix lint

* fix UTs

---------

Co-authored-by: Timothy J. Raymond <traymond@microsoft.com>